MANTA-5424 GC should be able to ignore storage ids that are unresolvable (#73)
This change makes it so that whenever a storageId is unreachable 10 consecutive times, (determined by PUT requests being slow and failing) garbage-uploader will stop trying to send further requests to that storageId for a period of time. This allows the system as a whole to continue to make progress in the face of a small number of servers that are unavailable.

Previously we'd be trying every file for these "down" servers which could be a large number if the server had been unreachable for a while, and the uploader would be stuck waiting for all those to time out on each run for a long time which caused delays in deletion for other servers which were not having problems.