MANTA-5427 garbage-dir-consumer needs to be able to handle large manta_fastdelete_queue tables (#74)
This changes garbage-dir-consumer to use .sql() and a manual query instead of findObjects when talking to moray. This is because findObjects will be slow whenever the table grows too large. Since findObjects normally also populates _count which we'd otherwise lose here, we've also implemented another query that runs at most once a minute to update the statistics for outstanding garbage.